[BEAM-1540] Move coders package classes to the top level
